modificari Model si Presentation - validari intrari, utilizare metode din ConsoleUtil etc. !!!! Mai am de actualizat StockMenu,Reader si Writer din Presentation, pachetul Service si pachetul Storage !!!

+++ src/main/java/ro/sda/dealership/presentation/CarMenu.java
@@ -2,21 +2,22 @@ package ro.sda.dealership.presentation;
 
 import ro.sda.dealership.model.Car;
 import ro.sda.dealership.storage.CarDAO;
+
+import java.util.InputMismatchException;
 import java.util.Scanner;
 
 public class CarMenu extends AbstractMenu {
-
     private CarDAO carDAO = new CarDAO();
     private CarReader reader = new CarReader();
     private CarWriter writer = new CarWriter();
 
     protected void displayOption() {
-        System.out.println("1.View all cars");
-        System.out.println("2.View car details");
-        System.out.println("3.Edit car");
-        System.out.println("4.Add new car");
-        System.out.println("5.Delete car");
-        System.out.println("0.Exit");
+        System.out.println("1. View all cars");
+        System.out.println("2. View car details");
+        System.out.println("3. Edit car");
+        System.out.println("4. Add new car");
+        System.out.println("5. Delete car");
+        System.out.println("0. Exit");
     }
 
     protected void executeComand(Integer option) {
@@ -25,19 +26,41 @@ public class CarMenu extends AbstractMenu {
                 writer.writeAll(carDAO.findAll());
                 break;
             case 2:
-               displayCarDetails();
+                if (carDAO.findAll().isEmpty()) {
+                    System.out.println("No cars available.");
+                } else {
+                    writer.writeAll(carDAO.findAll());
+                    System.out.print("Select car to view: ");
+                    displayCarDetails();
+                }
                 break;
             case 3:
-               editCarModel();
+                if (carDAO.findAll().isEmpty()) {
+                    System.out.println("No cars available.");
+                } else {
+                    writer.writeAll(carDAO.findAll());
+                    System.out.print("Select car to edit: ");
+                    editCar();
+                }
                 break;
             case 4:
                Car newCar = reader.read();
                carDAO.add(newCar);
                 break;
             case 5:
-                System.out.println("Select car to delete");
-                Long id = new Scanner(System.in).nextLong();
-                carDAO.deleteById(id);
+                if (carDAO.findAll().isEmpty()) {
+                    System.out.println("No cars available.");
+                } else {
+                    writer.writeAll(carDAO.findAll());
+                    System.out.print("Select car to delete: ");
+                    Long id = ConsoleUtil.readLong("Car");
+                    boolean isDeleted = carDAO.deleteById(id);
+                    if (!isDeleted) {
+                        System.out.println("Product not found");
+                    } else {
+                        System.out.println("Product deleted");
+                    }
+                }
             case 0:
                 System.out.println("Exiting to Main menu");
                 break;
@@ -46,21 +69,38 @@ public class CarMenu extends AbstractMenu {
         }
     }
 
-    private void editCarModel(){
+    private void editCar(){
         System.out.println("Select car to edit");
-        Long id = new Scanner(System.in).nextLong();
-        System.out.println("Enter new car: ");
-        String carModel = new Scanner(System.in).nextLine();
-        Car car = carDAO.findById(id);
-        car.setCarModel(carModel);
-        carDAO.update(car);
+        Scanner scanner = new Scanner(System.in);
+        Long id = ConsoleUtil.readLong("Car");
+        Car foundCar = carDAO.findById(id);
+        if (foundCar == null) {
+            System.out.println("Car not found");
+        } else {
+            System.out.println("Enter new model: ");
+            foundCar.setCarModel(scanner.next());
+            System.out.println("Enter new color: ");
+            foundCar.setCarColor(scanner.next());
+            System.out.print("Enter new price: ");
+            try {
+                foundCar.setPriceCar(scanner.nextDouble());
+            } catch (InputMismatchException e) {
+                System.out.println("Price not changed");
+            }
+            carDAO.update(foundCar);
+            System.out.println("Car updated");
+        }
     }
 
     private void displayCarDetails(){
         System.out.println("Choose car by Id: ");
-        Scanner scanner = new Scanner(System.in);
-        Long id = scanner.nextLong();
+        Long id = ConsoleUtil.readLong("Car");
         Car searchedCar = carDAO.findById(id);
-        writer.write(searchedCar);
+        if (searchedCar == null) {
+            System.out.println("Car not found");
+        } else {
+            System.out.println("Car details: ");
+            writer.write(searchedCar);
+        }
     }
 }
